allow supervisor to start services as pkg_svc_user
