+++ components/sup/src/manager/service/package.rs
@@ -18,12 +18,16 @@ use std::ops::Deref;
 use std::path::{Path, PathBuf};
 
 use hcore::fs::FS_ROOT_PATH;
+use hcore::os::users;
 use hcore::package::{PackageIdent, PackageInstall};
 use hcore::util::{deserialize_using_from_str, serialize_using_to_string};
 
 use error::{Error, Result};
 use fs;
 use util;
+
+const DEFAULT_USER: &'static str = "hab";
+const DEFAULT_GROUP: &'static str = "hab";
 
 const PATH_KEY: &'static str = "PATH";
 static LOGKEY: &'static str = "PK";
@@ -109,7 +113,7 @@ pub struct Pkg {
 
 impl Pkg {
     pub fn from_install(package: PackageInstall) -> Result<Self> {
-        let (svc_user, svc_group) = util::users::get_user_and_group(&package)?;
+        let (svc_user, svc_group) = get_user_and_group(&package)?;
         let pkg = Pkg {
             svc_path: fs::svc_path(&package.ident.name),
             svc_config_path: fs::svc_config_path(&package.ident.name),
@@ -143,5 +147,59 @@ impl Pkg {
             ),
         };
         Ok(pkg)
+    }
+}
+
+/// check and see if a user/group is specified in package metadata.
+/// if not, we'll try and use hab/hab.
+/// If hab/hab doesn't exist, try to use (current username, current group).
+/// If that doesn't work, then give up.
+fn get_user_and_group(pkg_install: &PackageInstall) -> Result<(String, String)> {
+    if let Some((user, group)) = get_pkg_user_and_group(&pkg_install)? {
+        Ok((user, group))
+    } else {
+        let defaults = default_user_and_group()?;
+        Ok(defaults)
+    }
+}
+
+/// This function checks to see if a custom SVC_USER and SVC_GROUP has
+/// been specified as part of the package metadata.
+/// If pkg_svc_user and pkg_svc_group have NOT been defined, return None.
+fn get_pkg_user_and_group(pkg_install: &PackageInstall) -> Result<Option<(String, String)>> {
+    let svc_user = pkg_install.svc_user()?;
+    let svc_group = pkg_install.svc_group()?;
+    match (svc_user, svc_group) {
+        (Some(user), Some(group)) => Ok(Some((user, group))),
+        _ => {
+            debug!("User/group not specified in package, running with default");
+            Ok(None)
+        }
+    }
+}
+
+/// checks to see if hab/hab exists, if not, fall back to
+/// current user/group. If that fails, then return an error.
+fn default_user_and_group() -> Result<(String, String)> {
+    let uid = users::get_uid_by_name(DEFAULT_USER);
+    let gid = users::get_gid_by_name(DEFAULT_GROUP);
+    match (uid, gid) {
+        (Some(_), Some(_)) => return Ok((DEFAULT_USER.to_string(), DEFAULT_GROUP.to_string())),
+        _ => {
+            debug!("hab:hab does NOT exist");
+            let user = users::get_current_username();
+            let group = users::get_current_groupname();
+            match (user, group) {
+                (Some(user), Some(group)) => {
+                    debug!("Running as {}/{}", user, group);
+                    return Ok((user, group));
+                }
+                _ => {
+                    return Err(sup_error!(Error::Permissions(
+                        "Can't determine current user:group".to_string(),
+                    )))
+                }
+            }
+        }
     }
 }
