+++ components/sup/src/util/mod.rs
@@ -16,4 +16,3 @@ pub mod convert;
 pub mod exec;
 pub mod path;
 pub mod pkg;
-pub mod users;
